refact: YearMonthPicker to fit new props
